ya se agregó y configuró la auntenticación por JWT al usuario, falta conectarlo con el frontend

+++ src/main/java/pe/sia/persistence/repository/ubicaciones/UbicacionRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import pe.sia.persistence.entity.ubicaciones.Ubicacion;
 
 @Repository
-public interface UbicacionRepository extends JpaRepository<Ubicacion, Long> {
+public interface UbicacionRepository extends JpaRepository<Ubicacion, Integer> {
 
 }
